Move PriorState components to their own files.
