+++ protocol-designer/src/pages/Designer/ProtocolSteps/StepForm/StepTools/HeaterShakerTools/PriorHeaterShakerState.tsx
@@ -1,0 +1,33 @@
+import type { HeaterShakerModuleState as SG_HeaterShakerModuleState } from '@opentrons/step-generation';
+import { useTranslation } from 'react-i18next';
+import { StepFormStatusList, StepFormStatus } from '/protocol-designer/components/molecules';
+
+export function PriorHeaterShakerState(props: {
+  priorState: SG_HeaterShakerModuleState;
+}): JSX.Element {
+  const { targetTemp, targetSpeed, latchOpen } = props.priorState;
+  const { t } = useTranslation();
+  return (
+    <StepFormStatusList>
+      <StepFormStatus
+        label={t('protocol_steps:heater_shaker.prior_state.heater_label')}
+        value={targetTemp != null
+          ? t('protocol_steps:heater_shaker.prior_state.heater_value', {
+            value: targetTemp,
+          })
+          : t('protocol_steps:heater_shaker.prior_state.heater_value_off')} />
+      <StepFormStatus
+        label={t('protocol_steps:heater_shaker.prior_state.shaker_label')}
+        value={targetSpeed != null
+          ? t('protocol_steps:heater_shaker.prior_state.shaker_value', {
+            value: targetSpeed,
+          })
+          : t('protocol_steps:heater_shaker.prior_state.shaker_value_off')} />
+      <StepFormStatus
+        label={t('protocol_steps:heater_shaker.prior_state.latch_label')}
+        value={latchOpen ?? false
+          ? t('protocol_steps:heater_shaker.prior_state.latch_value_open')
+          : t('protocol_steps:heater_shaker.prior_state.latch_value_closed')} />
+    </StepFormStatusList>
+  );
+}
