+++ protocol-designer/src/pages/Designer/ProtocolSteps/StepForm/StepTools/TemperatureTools/PriorTemperatureState.tsx
@@ -1,0 +1,26 @@
+import type { TemperatureModuleState as SG_TemperatureModuleState } from '@opentrons/step-generation';
+import { useTranslation } from 'react-i18next';
+import { StepFormStatusList, StepFormStatus } from '/protocol-designer/components/molecules';
+
+export function PriorTemperatureState(props: {
+  priorState: SG_TemperatureModuleState;
+}): JSX.Element {
+  const { t } = useTranslation();
+  const { targetTemperature } = props.priorState;
+  return (
+    <StepFormStatusList>
+      <StepFormStatus
+        label={t(
+          'protocol_steps:temperature_module.prior_state.temperature_label'
+        )}
+        value={targetTemperature != null
+          ? t(
+            'protocol_steps:temperature_module.prior_state.temperature_value',
+            { value: targetTemperature }
+          )
+          : t(
+            'protocol_steps:temperature_module.prior_state.temperature_value_off'
+          )} />
+    </StepFormStatusList>
+  );
+}
